Deposito v.3

+++ eclipse-workspace/Deposito/src/cuentas/CCuenta.java
@@ -1,7 +1,17 @@
 package cuentas;
+/**
+ * 
+ * @author Walter.Camacho
+ * Clase que nos da información sobre el estado de una cuenta de un cliente.
+ * @since 09-01-2020
+ * @version 1.1
+ * 
+ */
+
 public class CCuenta {
 
-    
+ // Variables con los datos de la clase. Nombre, cuenta, saldo y tipo de interés;
+	
     private String nombre;
     private String cuenta;
     private double saldo;
@@ -10,18 +20,37 @@ public class CCuenta {
     public CCuenta()
     {
     }
+    
+    /**
+     * 
+     * @param nom Pasa los datos del nombre del cliente
+     * @param cue Pasa el número de la cuenta
+     * @param sal Pasa el saldo
+     * @param tipo Pasa el interés.
+     */
 
-    public CCuenta(String nom, String cue, double sal, double tipo)
+    public CCuenta(String nom, String cue, double sal, float tipo)
     {
         setNombre(nom);
         setCuenta(cue);
         setSaldo(sal);
     }
+    
+    /**
+     * 
+     * @return el saldo actualizado
+     */
 
     public double estado()
     {
         return getSaldo();
     }
+    
+    /**
+     * 
+     * @param cantidad Ingreso realizado por el cliente
+     * @throws Exception Error de cantidad negativa
+     */
 
     public void ingresar(double cantidad) throws Exception
     {
@@ -29,6 +58,13 @@ public class CCuenta {
             throw new Exception("No se puede ingresar una cantidad negativa");
         setSaldo(getSaldo() + cantidad);
     }
+    
+    /**
+     * 
+     * @param cantidad. Cantidad que el cliente quiere retirar
+     * @throws Exception Errores a la hora de retirar bien por no tener esa cantidad en cuenta
+     * o bien por ser una cantidad negativa.
+     */
 
     public void retirar(double cantidad) throws Exception
     {
@@ -38,7 +74,10 @@ public class CCuenta {
             throw new Exception ("No se hay suficiente saldo");
         setSaldo(getSaldo() - cantidad);
     }
-
+/**
+ * 
+ * @return metodos que retornan el nombre del cliente, número de cuenta, saldo actualizado e interes
+ */
 	protected String getNombre() {
 		return nombre;
 	}
